Добавляет модули для функции-генератора, так же  добавляет импорты и экспорты
