+++ js/generator-id/comment-message.js
@@ -1,0 +1,13 @@
+const createMessage = () =>
+  Array.from({ length: getRandomInteger(1, 2) }, () =>
+    getRandomArrayElements(COMMENTS_STROKES)
+  ).join(' ');
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomInteger(1, AVATAR_COUNT)}.svg`,
+  message: createMessage(),
+  name: getRandomArrayElements(NAMES),
+});
+
+export {createMessage, createComment};
