+++ js/generator-id/create-generator-index.js
@@ -1,0 +1,12 @@
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return () => {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+const generateCommentId = createIdGenerator();
+
+export {createIdGenerator, generateCommentId};
